refactor: GameManger 인터페이스 제거
추가 확장 가능성이 거의 없기에 interface 제거
